requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
djangorestframework==3.12.2
idna==2.10
pika==1.2.0
pkg-resources==0.0.0
psycopg2-binary==2.8.6
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4